Rename dataset_base_enums to dataset_enums
