+++ backend/dataall/modules/datasets/db/dataset_models.py
@@ -2,7 +2,7 @@ from sqlalchemy import Boolean, Column, String, Text, ForeignKey
 from sqlalchemy.dialects.postgresql import JSON, ARRAY
 from sqlalchemy.orm import query_expression
 from dataall.base.db import Base, Resource, utils
-from dataall.modules.datasets.services.datasets_base_enums import ConfidentialityClassification, Language
+from dataall.modules.datasets.services.datasets_enums import ConfidentialityClassification, Language
 
 
 class DatasetTableColumn(Resource, Base):
